Compiled everything with -Wall, probably got most vuid_t vs. uid_t issues by using uid_t all over venus and only passing vuid's to the servers.

+++ coda-src/venus/venusfid.h
@@ -80,7 +80,8 @@ inline char *FID_(const VenusFid *fid)
     static int i = 0;
     i = 1 - i;
     sprintf(buf[i], "%x.%x.%x.%x",
-	    fid->Realm, fid->Volume, fid->Vnode, fid->Unique);
+	    (unsigned int)fid->Realm, (unsigned int)fid->Volume,
+	    (unsigned int)fid->Vnode, (unsigned int)fid->Unique);
     return buf[i];
 }
